Continue cleanup for Anod attributes
Remove source_list, not needed in the Anod class.

Use @property for source_pkg_build and package since that's how Anod
specification files are written. Also there is often the need
to check qualifier (for package) or sandbox properties (for
source_pkg_build) so using a class attribute is not often possible.

Change-Id: I553b7f0f2a5be15bbb39ed712d6fb10e4b5e7ab0