Fix sync_tree bug involving symlink
If a source directory contains a regular file having a corresponding
symlink with the name same name in the dest directory then the
symlink needs to be deleted before trying to change its content.
Otherwise the content of the symlink target will be modified and the
symlink will be untouched.

Also add an associated regression test.

Change-Id: Ieff3436eead6d36c868e45782b7f4944dbd64635
TN: P720-004