Fix e3.fs.rm when no exec permission on a dir to delete
Found when writing a test to cover e3.fs.rm code.

When there is no exec permission on a dir to delete, shutil.rmtree raise
an error when running os.listdir() and abort the deletion of the
subdirectory.  To continue deleting the directory content, e3.fs.rm now
add the missing permission to the subdirectory and then perform a
recursive delete of the directory.

Add test.

Change-Id: I580a15eee4a9ab9ec9df86da03f80ceffde0c7a5