Add new write_local_diff function in e3.vcs.git
Use diff-tree to dump the commit diff instead of git diff COMMIT^! which
does not work for the inital commit.

Use of closing "contextlib" in parse_log to make sure that we're not
relying on the reference counting to close the file. Also make
computation of the file size more reliable.

Add tests.