new (abstract) classes JobController and StatusBasedJobController in e3.jobs
This patch first introduces the root class JobController providing
abstract methods get_job and collect. The interface of those two
methods is intentionally set to match the "get_job" and "collect"
parameters of e3.job.Scheduler.__init__.

It then introduces a class StatusBasedJobController deriving from
JobController, which is still abstract, but specializes it to
provide handling for the job status of the job we create as well
as the status of its predecessors.

For QB17-039.