add "status" property to class Job
Add a comment explaning why this is introduced as a property rather
than an attribute.  Make it an abstractproperty to make sure child
classe always remember to define it.

Adjust all child classes as well as users of class Job accordingly.
Then simplify a bit the implementation of StatusBasedJobController.collect.

For QC20-017.