change Fingerprint.elements to a simple dict rather than an OrderedDict
There is no need for the elements of a fingerprint to be stored in
an OrderedDict, so this patch changes the implementation to use
a dictionary instead, taking care to sure that the behavior remains
independent of the order the elements are being added. Additional
testing is added to verify this property.

This will also simplify the implementation of methods for saving
and restoring fingerprints to/from files.

TN: QB17-039