Remove e3.anod.fingerprint.Fingerprint.add_instance
This is preparation work for moving the Fingerprint class outside
of Anod, as the concept of fingerprint is not specific to Anod.
Later on, we can either create an Anod Fingerprint class deriving
from the e3 Fingerprint class, or we can simply make this a simple
procedure/function, depending on what makes better sense.

TN: QB17-039