class BuildSpace: use class Fingerprint's load_from_file/save_to_file
This patch simplifies the implementation of class BuildSpace by
implementing the load-from-file/save-to-file methods using class
Fingerprint's load_from_file and save_to_file methods.

This has a side-effect of changing the format the fingerprint is
saved as: Instead of being dumped in yaml format, the fingerprint
is now dumped in JSON.  While yaml is slightly easier to read by
a human, JSON is said to be much faster. Also, if manual reading
of a fingerprint is needed, it's easy to issue a few Python commands
on the fly to read its contents and dump it in any other format.

To reflect this change of format, the fingerprint filename's extension
is changed to .json.

TN: QB17-039