add data parameter to e3.job.walk.Walk::compute_new_fingerprint
As it happens, the uid is not sufficient to compute a fingerprint,
we might also need the data. This makes sense, since we also pass
the data when creating the job, meaning that the data can influence
the job we create. So it should also be taken into account when
creating the corresponding fingerprint.

TN: QB17-039.