e3.job.walk.Walk: Add support for computing fingerprint after action done
There are some cases where we cannot compute the fingerprint ahead of
running a given action in the plan (eg: the fingerprint of an action
that contains the result of a download, for instance). So this patch
enhances the Walk class by adding a new method to class Walk called
can_predict_new_fingerprint. Child classes can override it to return
True/False based on the job.

The testsuite is extended to test that scenario.

TN: QB17-039.