Walk: better handling of actions that do not have a fingerprint
If actions in a plan cannot have a fingerprint computed for them,
then we should in this case always execute the actions that depend
directly on it. This patch enhances the should_execute_action method
to make sure we do.

A test is also added to verify this behavior.

TN: QB17-039.