robustify fingerprint save-to/load-from file
The purpose of this change is to make it easier to extend fingerprints
in the future while still being able to support older versions (if
needed). At the moment, it should make no difference in terms of
the amount of information stored, but better start using this earlier
rather than later.

This commit also adjust the load_from_file class method to return
None instead of raising E3Error. This is allows us to handle this
kind of situation as if there was no fingerprint, and just rerun
the corresponding action. What we want to avoid is a user not handling
the exception and crashing, rather than just deleting the fingerprint
file and continue normally as if the file did not exist.

TN: QB17-039