New API version to move .yaml files in subdirs
Instead of doing complex operations to detect which yaml files
are associated with .anod files, move .yaml files to subdirectories.

So to load the yaml file with the suffix 'suffix' associated with
the spec 'spec', look at 'spec/suffix.yaml'.

Change the default when the suffix is empty to look at
'spec/config.yaml'

Also set the api_version to the default version in AnodSpecsRepository
instead of requiring the user to set it in the prolog.

TN: R507-009