e3.job.walk: fix crash after failed predecessor with no predicted fingerprint
This patch fixes a crash while computing the fingerprint of a job
which depends on a job for which we could not compute the fingerprint
in advance and then failed. The crash was caused by the fact that
the new_fingerprints hash was missing an entry for the job
we depended on.

TN: R820-028 (crash report)
TN: QB17-039 (main e3 fingerprint work TN).