Greatly improve e3-plan-checker performance
Around 30% of the time spent when running e3-plan-checker is spent in
calls to relpath. These calls are done only to compute the ``ignore``
list that is used when installing source package content.

Instead of computing the ``ignore`` attribute when adding an entry in
the DAG move the computation in a property that will be called only when
needed, so only when installing source packages.

TN: R907-018