Limit leaks of file descriptors from collect and job_provider.
Calls to collect and job_provider are now protected to ensure that
no calls to Run is done by a job thread. This ensure fd opened by
collect and or job_provider are not leaked to these processes.

The general issue of file descriptor cannot be addressed in python
2.7.x. Some solution will be available once support for python 2.x
is dropped. See PEP 433, 446