factorize __init__ of AnodAction clas hierarchy
Note that this patch introduces a small API incompatibility when
calling AnodAction.__init__, as this patch changes the name of
one of the parameters from "data" to "anod_instance". This should
be very minor, as users of this code are expected to be calling
the __init__ of child classes instead, so we do not expect much
of an impact on existing code.

TN: R416-031