Parse source_dep of source dependencies in force_source_deps
Since the process we're currently spawning both generate sources and run
the build/test primitives, we also need to propagate the dependencies of
their source_pkg deps.

e.g. if for building A we need a source package defined in B, then we
need to add the source_deps of B when building A.

TN: RC10-034