Create a separate instance for each source we create
When loading a plan containing a "source" action, a separate run
of "anod source" if done for each source package. In order to avoid
side effect it is better to create a new instance for each source
package. Modify the cache key to take source_name into account.

TN: S905-021