Fix wrong error message in case of plan conflicts
First error detected by mypy: Decision.description takes either
Decision.LEFT or Decision.RIGHT as a parameter and we were passing
a string. The result is that we were always writing the same
error message, which was half of the time misleading.

Add type annotation, even if not perfect, to improve the detection
of similar mistakes in the future.