Sanitize Env.set_env input to handle bad user input
When a user pass the same value for build and host (or host and target)
Env.set_env currently creates different Platform objects and then
checks such as:

    if env.build != env.host: ...

have a surprising behaviour for users.

Let's handle this case in set_env, which is called when reading a
plan.

TN: R502-041