Refactor log arguments and configuration.
To make using e3.log possible without e3.main, I've pulled the
command-line arguments for the logging group out of e3.main and put them
in e3.log. I've also created a new activate method that is driven by the
parsed arguments.

To make e3.log a bit more general, I allow the default logging level to
be specified. The default level is taken to be WARNING; in e3.main, this
is overridden to INFO, to match prior behavior.

I also downcased the titles of the argument groups, as this seems more
consistent with the output of argparse.

Finally, I cleaned up the language on one of the help options for the
log.