Fix issue #417: replace `os.rename`
Instances of `os.rename` in e3 are replaced to use `e3.fs.mv`, which was
itself written to replace `shutil.move`.

Note that in `e3.fs.mv`, the uses of `os.rename` are retained. Quick
testing suggests that these uses do not cause problems, and the error
behavior of `os.rename` is sufficiently different from `shutil.move`
that I am reluctant to change these uses.

In the test case for `e3.fs`, the use of `os.rename` is replaced with
`shutil.move` - it seemed self-serving to use `e3.fs.mv` here.

Finally, in fix-coverage-paths, the use of `os.rename` is replaced with
`shutil.move`. This script doesn't import e3 at all, so sticking with
python libraries seemed best.