AnodContext: disable checks on DAG reversal
The "reverse_graph" DAG method currently checks that there are no cycle
in the *original* DAG (not in the returned reversed one). Make it
possible to disable this check, which is costly, and disable them in
AnodContext. This optimization removes 11% of the time spent in plan
checking in the most common case.

TN: TB27-006