small clarification about Run.status attribute
A new user of that class read the documentation and got confused
about the availability of the "status" attribute, thinking that
users must always call Run.wait for it to become available.
That's only the case when the object is initialized with bg=False,
so this commit expands the documentation a bit to make that clearer.

TN: V107-004