conftest.py: alter exit code on error
As described in the ticket, we need to be able to differentiate between
test failures (which are "normal" and should show up in the test report)
and legitimate errors encountered during setup/teardown.
A mechanism is put in place to detect those errors, and make pytest emit
a different exit code, so that we can detect that the testsuite
essentially failed to fully complete.

TN: V204-011