Improve the archive API with fileobj parameters
The archive API only allows unpacking an archive from file, and
writing an archive to file. However the zipfile, and tarfile
modules this API is based on, both accept fileobj parameters to
read from, and write to memory.

This change adds a fileobj parameter to create_archive, and
unpack_archive, in order to allow creating archives to, and
unpacking archives from memory.