Amend Anod primitive signature.
The arguments "pre" and "post" of an Anod primitive used to be Callables.
The real use for those parameters is to give them a string (a primitive name).

In order fix this, the signature of the primitive has been modified to allow
using a string.

The behaviour has been fixed too. The post parameter was not used.

Tests have been added to make sure all cases are covered.