e3-pypi-closure: Add yanked management
As stated in PEP_592, a yanked package should not be installed unless it
is the only one matching a version specifier.

It used to be possible for e3-pypi-closure to choose a yanked package, if
it was simply the most recent version of the package.

With this patch, no yanked package will be installed without explicit use
of the --allow-yanked option.