Add save method to VEX document
The VEX documents may now be saved to a file, using the
`Document.save(path, output_format)` method. The possible output formats
are `json` and `yaml`.

A subcomponent status may also be optional now. This means that the
statement status is inherited by the subcomponent if the subcomponent's
status is `None`.

It is also possible to specify an url for the vulnerability class.