Fix pytest error
When pytest is used with xdist and junitxml, tests are not run in the main
thread. In addition, signals can only be used in the main thread.

This results in the failure of some tests that attempt to define a signal
handler.

The signal handler causing the error is not tested, so we simply disable
the handler if the current thread is not the main thread.