Fix program name in main.Main class
When creating a program using `main.Main`, the program name displayed
by the `--help` message is not set to the *name* parameter of the
`main.Main` class.

Unless an `argument_parser` is provided to `main.Main.__init__()`, the
name of the arguments parser is now set to the *name* parameter.