Make anod external checkout ignore only asked files
Before this patch, VCS_IGNORE_LIST was always ignored during an external
checkout, whereas it is not ignored during an svn/git checkout.

This causes problems for certain tests and also inconsistency between
different checkouts.

So we've standardised things a little better.

Note that ‘.git’ and ‘.svn’ are always ignored because they are
potentially heavy folders that we never want to have in our sources.

it/e3-core#5