Remove e3.config and typeguard dependency
The initial goal of e3.config was to have a structured e3.toml file
that could be use to configure the behaviour of e3-core. This was
using typeguard to ensure that the config file was valid. This code
has only been used in e3.log, to configure the default log format.

Since March 2023, the code is no longer compatible with recent
versions of typeguard. Given that it is unmaintained and no longer in
use it is better to remove the code.

for it/e3-core#6