Emit error if shared lib is not found
The `check_shared_libraries_closure()` method does not raise an
exception if one of the share libraries is not found.

As the goals here is to make sure the shared libraries link with
appropriate shared libraries, the check should also raise an error
if a `not found` message is returned by `lld`.

An error is now raised when the `not found` message is raised by one
of the shared libraries.

Closes #7