Improve pypi closure computation by using a complete resolver
Previous approach was using a naive algorithm that was failing to
resolve complex constraints, mainly because of the absence of
backtracking. In order to solve properly the issue, the new
implementation drags the generic constraint solver used by pip
(resolvelib) and instantiate it.

Note also that the package now only depends on the standard
PyPI simple API rather than the release obsolete one.