Fix Git_Cmd typing
Git_Cmd is defined as List[Optional[str]] as to allow for lists
of str with potentially None values. However as lists are invariant,
mypy complains when a variable of type List[str] is passed with
the message `has incompatible type "list[str]";
expected "list[Optional[str]]"`. So Git_Cmd should also be defined
as List[str]