Detect invalid plan containing nested entry points
Nested entry points are not supported. When there is an nested
entry point, it is currently added to the list of entry point
only after having executed the parent entry point. As a result, it is
ignored by most tools and the code in the child entry point is
never executed.

For it/e3-core#14