Fix tests using `pip wheel` without network access
To workaround the fact that we don't have network access when running
`pip wheel` we can pass the argument `--no-build-isolation` and have
`the tool retrieve the build dependency from the current environment.

It's a good idea to also pass `--no-index` to ensure that the tests
won't try to access the PyPI server.

for it/e3-core#15