Fix shared libraries closure test
Some tests are executed on the current interpreter, but with a
static list of ignored shared libraries. This list is very
host-specific, and fail on nightly GAIA tests.

The list of ignored shared libraries is now incomplete, an error should
be raised. The test now makes sure that none of the shared libraries
reported as errors belong to the list of ignored libraries.

Closes #16