Fix shared libraries closure logic
The logic to group linked libraries with files is broken when a single
file is passed to `ldd`.

In order to make sure that the linked libraries are associated to an
executable/library, the method now returns the association as analysed
by `ldd`.

If a single file is passed to `ldd`, a new test ensures that the
analysis happened.

Closes #17