Handle tarfile.extractall() issue
As reported by `bandit` B202 handler, the `tarfile.extractall()` method
is not safe with some file paths.

We previously ignored this handler, and decided to handle things in
the `e3.archive` module now. The  `tarfile.data_filter()` is now used
when it exists, and a fallback handling absolute paths and traversal
paths has been created.

Ref: https://github.com/AdaCore/e3-core/issues/694