Basic support for pathlib.Path in e3.fs
Given the increasing use of pathlib.Path, and given that accepting
Path instead of str is easy: support is already there in os.path and
many os function, it makes sense to introduce a basic support for
pathlib.Path.

Path are accepted and converted using the recommended os.fspath
function when needed.

for it/e3-core#20