Disable branch coverage for e3-core
This is required in combination with COVERAGE_CORE=sysmon to speed up
coverage runs with Python >= 3.12.
See https://github.com/nedbat/coveragepy/issues/1812

This was tentativly done by it/e3-core!76 as the option was not
found in the tox.ini template file.

ref it/cookiecutter-e3#73
ref it/e3-core#28

See merge request it/cookiecutter-e3!61 of the template