Add support for argument checker in CommandResult
CommandResult accepts a list of str to be checked against the
actual arguments of the command. It often conflicts with
absolute paths pointing to the user directory or to a temporary
directory. This commit adds an ArgumentChecker protocol that can
be used with CommandResult to for example check if an argument
of the command matches a glob and then avoid such conflicts