Fix maven request
Solsearch API (The one used to retrieve available packages for Maven) is
paginated.

Previously, the maximum number of packages retrieved was 20. However, this
is not enough and not reliable because it is hardcoded.

This MR fix this issue by first making a request to retrieve the number
of packages and then the real request to get all related data.

Doing things this way may cause another issue because the request may be
huge if the number of packages to retrieve is high. But, this case never
occurs in practice, so it should be OK for our current needs.

Part of it/package-registry#27