MavenLink: Remove exception when checksum is not present
Having checksum is not a garantee for all package links. Raising exception
in that case seems not justify and its more the code that use MavenLink
that should manage error cases.

Indeed, with the exception raised, calling Maven().fetch_project_links on
a package with a single file without checksum prevent us to get all other
files.

Part of it/package-registry#27