Error with sync_tree when source top is a symlink
The call to `sync_tree` should be doing the same as
`rsync -rlpgoD --update`. When the top-level source element is a
symbolic link, `rsync` creates a true folder as the target top-level
element, then synchronizes the content.

The `sync_tree` function create a symbolic link instead, which may lead
to errors when trying to copy symbolic links.

The `sync_tree` now uses the real path of the top-level source element,
preventing this kind of error.

Closes #45+