Fix qualifiers manager type annotation
The `QualifiersManager.declare_component()` type annotation for the
`required_qualifier_configuration` parameter is misleading.

For a key value, the way to set the value to `True` is to pass an
empty string `""` for value. This is misleading when using the `"False"`
string, the value is in fact set to `True`.

The parameter type annotation has been updated to allow boolean value.
This meant some other type annotation changes in the class to remain
coherent.

Ref: IT-2691