Shared knowledge: Add way to insert plan symbol in plan module
All plans are concatenated into a single Python module. As with all Python
modules, we can modify its `__dict__` and accessible symbols. This MR adds
a way to do this once the plan module is created.

This is useful for customizing the plan module and sharing knowledge
between Python modules. For example, we can add all loaded specs to a plan
to make their functions accessible from the plan.

For example, by injecting the `spec` symbol as `AnodSpecRepository.load`
you can create the spec `myspec.anod` like this:
```
class MySpec(Anod):
    @classmethod
    def my_default_build_param(cls) -> dict[str, object]:
        return {"target": "x86_64-linux"}
```

And then, use the class method in the plan, like this:
```
def my_server():
    with defaults(**spec("myspec").my_default_param()):
        anod_build("myspec")
```
Doing so, `myspec.anod` will be built using `target="x86_64-linux"`.

it/e3-adacore#85