Fix use of invalid parameter
In commit c021846c7c62248771b26ef4061836d3352ce0f7, the call to
`os.path.realpath` uses the `strict` parameter.

Unfortunately, this parameter does not exist for python versions
prior to 3.10. The default python version on GitHub CIs being 3.9, this
parameter is problematic.

This parameter is not used anymore for python versions prior to 3.10.

Ref https://github.com/AdaCore/e3-core/issues/738