Fix e3.python.pypi.PyPI.fetch_candidates
If an error is generated during the link parsing, no candidates are
returned when trying to fetch `PyPI` candidates for a given package.

For example, an error will be generated when parsing `bson` links due to a
non-standard archive name for `bson 0.4.5` (`bson__0.4.5.tar`).

This commit enables the `ignore_errors` flag when parsing `PyPI` links,
allowing us to not consider packages with issues as candidates without
generating an error.

it/e3-distrib#7