Change some Store tables id type
Managing string instead of int as table ID is more scalable: we always
have a way to transform an object into a string, but that is not the case
for an int.

Additionally, we need that scalability to implement a local interface
required for the project's next step.

This commit also fixes the release retrieval for a given component. The
code used the wrong table field, resulting in a false result.

it/org/software-supply-chain/production-pipeline/issues#1