Remove empty logs from rm command
The function logs some debug information on the files it is about to
delete.

The problem described in #56+ is that this log happens even if there are
no file(s) to remove.

The debug log has been removed when there are no file(s) to remove.

A previous attempt to fix this issue was made; however, the `rm` function
sometimes got a file list like: `file_list=[""]` resulting in a non-empty
`file_list`, but trying `" ".join(file_list)` still results in an empty
string.

Closes #56+
Requested by IT-2956