Improve e3.anod.store.Store.get_source_info
When looking for source files (kind="source"), the build ID is not a
strict match. A source file should be re-generated only if the previous
source has changed, and so, a source file will rarely match exactly today's
build ID.

This commit changes how we look for a file. In the case of sources, the new
code will return the latest source file with linked to a buildinfo OLDER
than the one referenced.

Related to it/org/software-supply-chain/production-pipeline/issues#1