Add component metadata
It is possible to add (extra) metadata at the Store file level through the
`set_metadata_statement()` and the `get_metadata_statement()` methods.

This helps, for instance, to save SBOM metadata for each source file.

Unfortunately, this does not exist at the component level. This commit
adds the possibility to add such metadata at a component level.

Related to it/org/software-supply-chain/production-pipeline/issues#1