Fix latest_component method
This method must return a list of components, but each component in this
list must have a unique name/platform tuple.

With the previous implementation, this function returned all components
matching the given criteria, but with several components having the same
name/platform tuple.

Related to it/org/software-supply-chain/production-pipeline/issues#1