Raise an HTTPError when NVD request failed
When using the `e3.cve.NVD.search_by_cpe_name()` method, if the NVD
site has a problem we get a JSON decode error, and the real error is not
raised.

Throwing the real `HTTPError` rather than a `RequestsJSONDecodeError`
would help knowing what the underlying problem is.

The real `HTTPError` is now raised through a call to
`Response.raise_for_status()` when the http query failed.

Closes #57+