Fix testing
`e3.sys.interpreter` can return `python3` or `python`. We have a test that
failed on Windows because `sys.executable` is sees as `python`, whereas
`e3.sys.interpreter` returns `python3`.

To fix this issue, we just remove the `3` for the test comparison, since in
any case, we are not supporting other Python than Python3.

it/e3-core#60