Fix documentation build
The issue comes from the astroid version used. Currently, `astroid>=4`
cannot be used to build the documentation.

This commit fix the documentation build on readthedocs by adding
`astroid>=3,<4` to the `docs/requirements.txt`.

The fix for `tox -e docs` will be done in the cookiecutter repository as
the `tox.ini` file is managed by cookiecutter. See: it/cookiecutter-e3!92

it/e3-core#61