Fix sphinx build error when generating documentation for e3.anod.qualifier
This commit fixes the following errors:
/docs/source/autoapi/anod/qualifier/index.rst:25: ERROR: Unexpected indentation.
/docs/source/autoapi/anod/qualifier/index.rst:26: WARNING: Block quote ends without a blank line; unexpected unindent.
/docs/source/autoapi/anod/qualifier/index.rst:29: ERROR: Unexpected indentation.
/docs/source/autoapi/anod/qualifier/index.rst:32: WARNING: Inline strong start-string without end-string.

This error is caused by the Python default documentation of a dictionary.
On Sphinx when no class documentation is provided, the default class
documentation becomes the documentation of the inherited type. Visibly, the
default documentation of the Python `dict` is not correct from the Sphinx
point of view.

This commit also improves some minor parts of the documentation of this
module.