Fix tests_e3:anod:store:component_test:py--test_component_metadata
The `creation_date` field, used to retrieve the latest component, has a
precision of seconds.

When two consecutive component pushes are done too quickly, it may result
in two components with the same `creation_date` field.

To avoid this error and make a better fix, we will simply sleep for one
second between two pushes.

it/e3-core#62