Remove a debug log in e3.fs.rm function
When the `shutil.rmtree()` function fails in the `e3.fs.rm` function,
we enter the `onerror()` function, which immediately emits a debug log
about this failure.

Our build logs on Windows are polluted by this log, and thousands of
lines about `unlink` error.

The debug log has been removed, but a debug log is kept if the file
could NOT actually be removed.

Closes it/e3-core#65+