Disable the download progress bar according to log level
As described in https://github.com/AdaCore/e3-core/issues/651, it
is not possible to disable the progress bar when downloading files
through the `e3.net.http.HTTPSession.download_file()` method.

In some cases, it may be boring, especially when within another progress
bar.

Assuming that, when a user sets the log level to `WARNING` and above,
he does not want to have the download progress bar to appear on his
log, the *disable* parameter of `e3.log.progress_bar()` call in
`e3.net.http.HTTPSession.download_file()` has been set according to the
log level.

Closes https://github.com/AdaCore/e3-core/issues/651