Add `component_id` field in the components table
This commit changes the meaning of the `id` field for the `components`
table. This new field will be used to order any request to get the last
element added more precisely. Bigger is the ID, more recent is the entry
in the database.

We made this because comparing the `creation_date` may causes issue
depending of the datetime precision and the version of SQLite used by
python. For exemple, with an old SQLite, the precision is in seconds, and
if two entries is added within the same second, the `creation_date`
comparison is not enougth to get the lastest inserted element.