Order request by ID instead of creation_date
This commit orders any request by the id field to get the last element
added to the database more precisely. The larger the ID, the more recent
the entry is in the database.

We made this because comparing the `creation_date` may cause issues
depending on the datetime precision and the version of SQLite used by
Python. For example, with an old SQLite, the precision is in seconds, and
if two entries are added within the same second, the `creation_date`
comparison is insufficient to determine the latest inserted element.