Fix a Component comparison issue
This commit normalize the `Component.creation_date` to reduce the datetime
precision. The datetime may have a bigger precision than SQLite; This
cause a problem when comparing two component, that must be equal, because
one of them may have a microsecond precision when the other may have a
milliseconds precision. To avoid that, when we are creating the
`creation_date` attribute (when not provided), this commit will remove
the microseconds from the object.